pyproject.toml
[project]
name = "weather-site"
version = "0.1.0"
description = "weather.gov, but better."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "duckdb",
    "dlt",
    "fastapi",
    "httpx",
    "pydantic",
    "python-dotenv",
    "uvicorn[standard]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/fastapi_app"]

[tool.uv]
package = true
